requirements.txt
PyQt5==5.15.4
pdfminer.six
